requirements.txt
numpy==1.26.0
opencv_python==4.8.1.78
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
